backend/requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
python-multipart==0.0.6
torch==2.0.1
torchvision==0.15.2
timm==0.9.5
albumentations==1.3.1
numpy==1.23.5
pillow==9.5.0
scikit-learn==1.3.0
pydantic==1.10.8
tqdm==4.66.1
opencv-python==4.8.0.76 